Compress the newly modified temporary directory to .tgz
This involved moving the temporary extracted directory outside of the created temporary directory, into a new directory, then compressing the moved files. (In order to preserve the desired directory structure). The newly created temporary folder is then supposed to be removed, but it doesn't currently. (Maybe due to the file permissions?)